Make Venmo Button field optional with default value, put Bundle extension into separate file
